Minimally update the README now that the repo is public (#205)
* Tweak blurb; fix docker instructions

* Minor tweaks

* Fix readme docker command

+++ scripts/init.sh
@@ -2,10 +2,14 @@
 
 set -o nounset
 
+echo -n "This script will remove all existing iovns configurations, wallets etc. Do you want to proceed? [Y/n] "
+read YN
+if ! echo $YN | grep -v -q -i "n"; then exit; fi
+
 rm -rf "$HOME/.iovnscli"
 rm -rf "$HOME/.iovnsd"
 # init config files
-iovnsd init "$MONIKER" --chain-id iovns
+iovnsd init "$(hostname)" --chain-id iovns
 # configure cli
 iovnscli config chain-id iovns
 iovnscli config output json
